Code tested and working

+++ 5-development/Data/preparedSession.py
@@ -1,14 +1,11 @@
-from dataclasses import dataclass, field
-from typing import List
-
-from matplotlib.pylab import double
+from dataclasses import dataclass
 
 @dataclass
 class PreparedSession:
     """A single sample: a feature vector and its label."""
     UUID: str = ""
     idPlayer: str = ""
-    skillOverall: double = 0.0
-    socialInfluence: double = 0.0
-    injuriesImpact: double = 0.0
+    skillOverall: float = 0.0
+    socialInfluence: float = 0.0
+    injuriesImpact: float = 0.0
     label: int = 0
